% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertJoToSquareMeter}
\alias{convertJoToSquareMeter}
\title{Convert unit from Japanese jo to metric.}
\usage{
convertJoToSquareMeter(jo)
}
\arguments{
\item{jo}{square measure in Japanese jo}
}
\value{
square_measure in metric
}
\description{
Convert unit from Japanese jo to metric.
}
