% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertSquareMeterToJapaneseJo}
\alias{convertSquareMeterToJapaneseJo}
\title{Convert unit from metric to Japanese jo.}
\usage{
convertSquareMeterToJapaneseJo(squareMeter)
}
\arguments{
\item{squareMeter}{square meter}
}
\value{
square_measure in Japanese jo
}
\description{
Convert unit from metric to Japanese jo.
}
