% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_tatami.R
\name{launch_tatami}
\alias{launch_tatami}
\alias{launch_tatami.default}
\title{Launch the 'Tatatmi' app}
\usage{
launch_tatami(first, ..., rstudio = getOption("rtatami.rstudio"))

\method{launch_tatami}{default}(first, ...,
  rstudio = getOption(("rtatami.rstudio")))
}
\description{
Launch the 'Tatami' app in the default web browser.
'RStudio' users also have the option of
launching the app in the pop-up Viewer.
}
